move definitions to cpp
