+++ example/tcp/CMakeLists.txt
@@ -6,7 +6,7 @@ add_executable(tcp_example
 )
 
 target_link_libraries(tcp_example
-	# tcp
+	tcp
 	pthread
 )
