+++ src/CMakeLists.txt
@@ -1,3 +1,4 @@
 
 add_subdirectory(process)
 add_subdirectory(logger)
+add_subdirectory(tcp)
